feat: don't use script's directory for storing files
Previously, embr used the base directory where embr was
located for storing downloaded images, vms, config, etc.

This was fine in early development, but for installs where
embr might be in /usr/bin or the nix store, this is not a
great plan.

This change introduces $HOME/.config/embr as the default root
directory for storing files and config.